moved rosetta scoring to its own file
