+++ src/nplinker/nplinker.py
@@ -27,7 +27,7 @@ from .pickler import save_pickled_data
 from .scoring.link_collection import LinkCollection
 from .scoring.methods import MetcalfScoring
 from .scoring.np_class_scoring import NPClassScoring
-from .scoring.methods import RosettaScoring
+from .scoring.rosetta_scoring import RosettaScoring
 from .scoring.methods import TestScoring
