+++ src/nplinker/scoring/methods.py
@@ -24,7 +24,6 @@ from nplinker.metabolomics import Spectrum
 from nplinker.pickler import load_pickled_data
 from nplinker.pickler import save_pickled_data
 from nplinker.scoring.object_link import ObjectLink
-from nplinker.scoring.rosetta.rosetta import Rosetta
 from .linking.data_linking import DataLinks
 from .linking.data_linking import LinkFinder
 
@@ -57,7 +56,6 @@ class ScoringMethod():
     def sort(self, objects, reverse=True):
         """Given a list of objects, return them sorted by link score"""
         return objects
-
 
 class TestScoring(ScoringMethod):
 
@@ -96,150 +94,6 @@ class TestScoring(ScoringMethod):
     def sort(self, objects, reverse=True):
         # nothing
         return objects
-
-
-class RosettaScoring(ScoringMethod):
-
-    NAME = 'rosetta'
-    ROSETTA_OBJ = None
-
-    def __init__(self, npl):
-        super().__init__(npl)
-        self.bgc_to_gcf = True
-
-        self.spec_score_cutoff = 0.0
-        self.bgc_score_cutoff = 0.0
-
-    @staticmethod
-    def setup(npl):
-        logger.info('RosettaScoring setup')
-        RosettaScoring.ROSETTA_OBJ = Rosetta(npl, ignore_genomic_cache=False)
-        ms1_tol = Rosetta.DEF_MS1_TOL
-        ms2_tol = Rosetta.DEF_MS2_TOL
-        score_thresh = Rosetta.DEF_SCORE_THRESH
-        min_match_peaks = Rosetta.DEF_MIN_MATCH_PEAKS
-
-        # allow overridding params via config file
-        config = npl.config
-        if 'scoring' in config and 'rosetta' in config['scoring']:
-            rc = config['scoring']['rosetta']
-            ms1_tol = rc.get('ms1_tol', Rosetta.DEF_MS1_TOL)
-            ms2_tol = rc.get('ms2_tol', Rosetta.DEF_MS2_TOL)
-            score_thresh = rc.get('score_thresh', Rosetta.DEF_SCORE_THRESH)
-            min_match_peaks = rc.get('min_match_peaks',
-                                     Rosetta.DEF_MIN_MATCH_PEAKS)
-
-        RosettaScoring.ROSETTA_OBJ.run(npl.spectra, npl.bgcs, ms1_tol, ms2_tol,
-                                       score_thresh, min_match_peaks)
-        logger.info('RosettaScoring setup completed')
-
-    def _include_hit(self, hit):
-        if hit.spec_match_score < self.spec_score_cutoff or hit.bgc_match_score < self.bgc_score_cutoff:
-            return False
-
-        return True
-
-    def _insert_result_gen(self, results, src, hit):
-        if src not in results:
-            results[src] = {}
-        # Rosetta can produce multiple "hits" per link, need to
-        # ensure the ObjectLink contains all the RosettaHit objects
-        # in these cases
-        if hit.spec in results[src]:
-            original_data = results[src][hit.spec].data(self)
-            results[src][hit.spec].set_data(self, original_data + [hit])
-        else:
-            results[src][hit.spec] = ObjectLink(src,
-                                                hit.spec,
-                                                self,
-                                                data=[hit])
-
-        return results
-
-    def _insert_result_met(self, results, spec, target, hit):
-        if spec not in results:
-            results[spec] = {}
-        # Rosetta can produce multiple "hits" per link, need to
-        # ensure the ObjectLink contains all the RosettaHit objects
-        # in these cases
-        if target in results[spec]:
-            original_data = results[spec][target].data(self)
-            results[spec][target].set_data(self, original_data + [hit])
-        else:
-            results[spec][target] = ObjectLink(spec, target, self, data=[hit])
-
-        return results
-
-    def get_links(self, objects, link_collection):
-        # enforce constraint that the list must contain a set of identically typed objects
-        if not all(isinstance(x, type(objects[0])) for x in objects):
-            raise Exception(
-                'RosettaScoring: uniformly-typed list of objects is required')
-
-        if isinstance(objects[0], MolecularFamily):
-            raise Exception(
-                'RosettaScoring requires input type Spectrum (found MolecularFamily)'
-            )
-
-        if isinstance(objects[0], GCF):
-            # assume user wants to use all BGCs from these GCFs
-            bgcs = list(
-                set(itertools.chain.from_iterable(x.bgcs for x in objects)))
-            logger.info(
-                'RosettaScoring got {} GCFs input, converted to {} BGCs'.
-                format(len(objects), len(bgcs)))
-            objects = bgcs
-
-        # list of RosettaHit objects which satisfy the current cutoffs
-        ro_hits = list(
-            filter(lambda hit: self._include_hit(hit),
-                   RosettaScoring.ROSETTA_OBJ._rosetta_hits))
-
-        # TODO this might need to be faster
-        results = {}
-        if isinstance(objects[0], BGC):
-            for bgc in objects:
-                for hit in ro_hits:
-                    if bgc.id == hit.bgc.id:
-                        if not self.bgc_to_gcf:
-                            # can use the BGC directly
-                            results = self._insert_result_gen(
-                                results, bgc, hit)
-                        else:
-                            # if we want the results to contain GCFs instead, need
-                            # to iterate over the set of bgc.parents (most will only
-                            # have one, hybrid BGCs will have more)
-                            for gcf in bgc.parents:
-                                results = self._insert_result_gen(
-                                    results, gcf, hit)
-        else:  # Spectrum
-            for spec in objects:
-                for hit in ro_hits:
-                    if spec.id == hit.spec.id:
-                        if not self.bgc_to_gcf:
-                            # can use the BGC directly
-                            results = self._insert_result_met(
-                                results, spec, hit.bgc, hit)
-                        else:
-                            # if we want the results to contain GCFs instead, need
-                            # to iterate over the set of bgc.parents (most will only
-                            # have one, hybrid BGCs will have more)
-                            for gcf in hit.bgc.parents:
-                                results = self._insert_result_met(
-                                    results, spec, gcf, hit)
-
-        link_collection._add_links_from_method(self, results)
-        logger.debug(f'RosettaScoring found {len(results)} results')
-        return link_collection
-
-    def format_data(self, data):
-        # TODO
-        return f'{len(data)} hits'
-
-    def sort(self, objects, reverse=True):
-        # TODO
-        return objects
-
 
 class MetcalfScoring(ScoringMethod):
